requirements.txt
playwright
pandas
google-genai
anthropic
python-dotenv
flask
